requirements.txt
cvxpy==1.3.0
cycler==0.11.0
ecos==2.0.12
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.5.3
mpmath==1.2.1
numpy==1.21.6
osqp==0.6.2.post8
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
qdldl==0.1.5.post3
scipy==1.7.3
scs==3.2.2
six==1.16.0
sympy==1.10.1
typing-extensions==4.4.0
